# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CitationOut": {
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "heading_path": {
            "title": "Heading Path",
            "type": "string"
          },
          "module_code": {
            "title": "Module Code",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "source_tier": {
            "enum": [
              "university",
              "own"
            ],
            "title": "Source Tier",
            "type": "string"
          }
        },
        "required": [
          "n",
          "document_id",
          "filename",
          "page_no",
          "source_tier",
          "module_code",
          "heading_path",
          "quote"
        ],
        "title": "CitationOut",
        "type": "object"
      },
      "ConversationCreate": {
        "additionalProperties": false,
        "description": "Start a conversation, optionally from a module or topic (searches\nstart there and widen when they find little).",
        "properties": {
          "module_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "title": "ConversationCreate",
        "type": "object"
      },
      "ConversationDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "title": "Messages",
            "type": "array"
          },
          "module_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "module_id",
          "topic_id",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ConversationDetail",
        "type": "object"
      },
      "ConversationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "module_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "module_id",
          "topic_id",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "MessageCreate",
        "type": "object"
      },
      "MessageLink": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "draft",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "label"
        ],
        "title": "MessageLink",
        "type": "object"
      },
      "MessageOut": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/PendingActionOut"
            },
            "title": "Actions",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/MessageLink"
            },
            "title": "Links",
            "type": "array"
          },
          "provenance": {
            "items": {
              "enum": [
                "university",
                "own",
                "general"
              ],
              "type": "string"
            },
            "title": "Provenance",
            "type": "array"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "enum": [
              "complete",
              "stopped",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "citations",
          "provenance",
          "steps",
          "status",
          "error_code",
          "created_at"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "PendingActionOut": {
        "properties": {
          "action": {
            "enum": [
              "delete_document",
              "delete_topic",
              "delete_module"
            ],
            "title": "Action",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "cancelled",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "preview",
          "status",
          "expires_at"
        ],
        "title": "PendingActionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Revision OS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "operationId": "list_conversations_api_v1_conversations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationOut"
                  },
                  "title": "Response List Conversations Api V1 Conversations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Conversations",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "operationId": "create_conversation_api_v1_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}": {
      "delete": {
        "operationId": "delete_conversation_api_v1_conversations__conversation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Conversation",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "operationId": "get_conversation_api_v1_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Conversation",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}/messages": {
      "post": {
        "description": "Ask a question; the answer streams back as server-sent events:\n\n- `status` {text}: what the assistant is doing (\"Searching \u2026\")\n- `delta` {text}: the next piece of answer text\n- `action` PendingAction: a deletion awaiting your confirmation\n- `done` Message: the saved answer, with verified citations\n- `error` {code, message, saved: Message}: the answer failed or was refused",
        "operationId": "send_message_api_v1_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/v1/pending-actions/{action_id}/cancel": {
      "post": {
        "operationId": "cancel_pending_action_api_v1_pending_actions__action_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Pending Action",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/pending-actions/{action_id}/confirm": {
      "post": {
        "description": "Carry out a deletion Claude requested. Only reachable from your signed-in\nbrowser with a CSRF token; the item goes to the trash.",
        "operationId": "confirm_pending_action_api_v1_pending_actions__action_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Pending Action",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}